scripts/iam/requirements.txt
appdirs==1.4.3
attrs==19.1.0
aws-mfa==0.0.12
black==18.9b0
boto3==1.9.107
botocore==1.12.107
Click==7.0
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==3.13
s3transfer==0.2.0
six==1.12.0
toml==0.10.0
urllib3==1.24.1
